support fallback ECDH for android v7 <7.1.1
The android app is using react-native for HTTP requests via the fetch
API. This is implemented with okhttp. It's not possible to create a
connection with okhttp to the web server because some android versions
(version 7 < 7.1.1) have an bug to use the ECDH curve ECP384.

To support this android versions we can configure a fallback curve in
nginx that should be strong enough: Prime 256.

For more information about the bug: https://issuetracker.google.com/issues/37122132